# openapi-order-module.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "营养餐订单系统 API",
    "description": "营养餐订单管理系统的API接口规范，包含地址管理、订单管理和配送管理",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API服务器"
    }
  ],
...
  "paths": {
...
    "/orders": {
      "get": {
        "tags": ["订单管理"],
        "summary": "分页查询订单",
        "description": "返回当前登录用户的订单列表，支持分页和按状态查询",
        "operationId": "pageOrders",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "订单状态：0-待支付，1-已支付待配送，2-配送中，3-已完成，4-已取消",
            "required": false,
            "schema": {
              "type": "integer",
              "enum": [0, 1, 2, 3, 4]
            }
          },
          {
            "name": "orderNo",
            "in": "query",
            "description": "订单编号",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageNum",
            "in": "query",
            "description": "页码，从1开始",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页数量",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取订单列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPageResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": ["订单管理"],
        "summary": "创建订单",
        "description": "创建新的订单",
        "operationId": "createOrder",
        "requestBody": {
          "description": "订单信息",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "订单创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderIdResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": ["订单管理"],
        "summary": "获取订单详情",
        "description": "根据订单ID获取订单详情",
        "operationId": "getOrderDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "订单ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取订单详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "订单不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/orders/{id}/cancel": {
      "post": {
        "tags": ["订单管理"],
        "summary": "取消订单",
        "description": "取消指定ID的订单",
        "operationId": "cancelOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "订单ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "订单取消成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "订单状态不允许取消",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "订单不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/orders/{id}/pay": {
      "post": {
        "tags": ["订单管理"],
        "summary": "支付订单",
        "description": "支付指定ID的订单",
        "operationId": "payOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "订单ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "订单支付成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "订单状态不允许支付",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "订单不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/orders/{id}/confirm": {
      "post": {
        "tags": ["订单管理"],
        "summary": "确认收货",
        "description": "确认指定ID的订单已收货",
        "operationId": "confirmReceive",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "订单ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "确认收货成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "订单状态不允许确认收货",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "订单不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "type": "null",
            "description": "无数据"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功",
            "default": false
          },
          "message": {
            "type": "string",
            "description": "错误信息"
          },
          "data": {
            "type": "null",
            "description": "无数据"
          }
        }
      },
      "OrderPageResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "description": "总记录数"
              },
              "pages": {
                "type": "integer",
                "description": "总页数"
              },
              "orders": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        }
      },
      "OrderRequest": {
        "type": "object",
        "properties": {
          "addressId": {
            "type": "integer",
            "format": "int64",
            "description": "地址ID"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "totalAmount": {
            "type": "number",
            "format": "double",
            "description": "总金额"
          },
          "payMethod": {
            "type": "string",
            "description": "支付方式"
          },
          "payTime": {
            "type": "string",
            "format": "date-time",
            "description": "支付时间"
          },
          "status": {
            "type": "string",
            "description": "订单状态"
          }
        }
      },
      "OrderIdResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "type": "integer",
            "format": "int64",
            "description": "订单ID"
          }
        }
      },
      "OrderDetailResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "$ref": "#/components/schemas/Order"
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "订单ID"
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "description": "用户ID"
          },
          "addressId": {
            "type": "integer",
            "format": "int64",
            "description": "地址ID"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "totalAmount": {
            "type": "number",
            "format": "double",
            "description": "总金额"
          },
          "payMethod": {
            "type": "string",
            "description": "支付方式"
          },
          "payTime": {
            "type": "string",
            "format": "date-time",
            "description": "支付时间"
          },
          "status": {
            "type": "string",
            "description": "订单状态"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "订单项ID"
          },
          "orderId": {
            "type": "integer",
            "format": "int64",
            "description": "订单ID"
          },
          "productId": {
            "type": "integer",
            "format": "int64",
            "description": "产品ID"
          },
          "quantity": {
            "type": "integer",
            "format": "int64",
            "description": "数量"
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "单价"
          },
          "totalAmount": {
            "type": "number",
            "format": "double",
            "description": "总金额"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
} 